Use a nn.Module as the top-level class.
