+++ mlx_audio/tts/models/sesame/__init__.py
@@ -1,3 +1,3 @@
-from .generator import Model
+from .model import Model
 
 __all__ = ["Model"]
